+++ src/main/controllers/withdrawStake.ts
@@ -48,12 +48,16 @@ const withdrawStake = async (
   try {
     mainSystemAccount = Keypair.fromSecretKey(
       Uint8Array.from(
-        JSON.parse(fsSync.readFileSync(mainWalletfilePath, 'utf-8'))
+        JSON.parse(
+          fsSync.readFileSync(mainWalletfilePath, 'utf-8')
+        ) as Uint8Array
       )
     );
     stakingAccKeypair = Keypair.fromSecretKey(
       Uint8Array.from(
-        JSON.parse(fsSync.readFileSync(stakingWalletfilePath, 'utf-8'))
+        JSON.parse(
+          fsSync.readFileSync(stakingWalletfilePath, 'utf-8')
+        ) as Uint8Array
       )
     );
   } catch (e: any) {
